tidy: size_t casts because windows

+++ tactless.c
@@ -599,7 +599,7 @@ static int parse_encoding(byte *s, size_t size, struct encoding *e) {
     return 0;
   }
   const byte *index = s + 22 + espec_block_size;
-  const byte *data = index + 32L * cekey_page_count;
+  const byte *data = index + (size_t)32 * cekey_page_count;
   int entries = 0;
   for (const byte *ic = index, *dc = data; ic != data; ic += 32, dc += 4096) {
     if (!md5check(dc, 4096, ic + 16)) {
@@ -619,7 +619,7 @@ static int parse_encoding(byte *s, size_t size, struct encoding *e) {
   if (!entries) {
     return 0;
   }
-  byte *arr = malloc(entries * 32L);
+  byte *arr = malloc(entries * (size_t)32);
   if (!arr) {
     return 0;
   }
